+++ .dockerignore
@@ -1,0 +1,4 @@
+target
+.idea/
+*.iml
+*.iws
