+++ docker-compose.dev.yml
@@ -1,0 +1,21 @@
+version: '3.8'
+services:
+  mysql-springboot-container:
+    image: mysql:8.0.28
+    container_name: mysql-docker-container
+    ports:
+      - "3310:3306"
+    networks:
+      - backend
+    environment:
+      MYSQL_ROOT_PASSWORD: Abc@123456789
+  springboot-docker-container:
+    image: pmt296/springboot-docker:v1.0.0
+    container_name: springboot-docker-container
+    ports:
+      - "8085:8083"
+    networks:
+      - backend
+networks:
+  backend:
+    name: springboot-app-network
