+++ src/main/java/com/pmt/tool/dto/TRoleDto.java
@@ -1,0 +1,10 @@
+package com.pmt.tool.dto;
+
+import lombok.Data;
+
+import java.io.Serializable;
+
+@Data
+public class TRoleDto implements Serializable {
+    private String nameRole;
+}
