+++ src/main/java/com/pmt/tool/dto/TUserDto.java
@@ -3,6 +3,7 @@ package com.pmt.tool.dto;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.util.List;
 
 @Data
 public class TUserDto implements Serializable {
@@ -11,4 +12,5 @@ public class TUserDto implements Serializable {
     private String firstName;
     private String lastName;
     private Boolean gender;
+    private List<TRoleDto> roleDtoList;
 }
