+++ src/main/java/com/pmt/tool/entity/TDetailWork.java
@@ -3,7 +3,6 @@ package com.pmt.tool.entity;
 import lombok.*;
 
 import javax.persistence.*;
-import java.io.Serializable;
 
 @AllArgsConstructor
 @NoArgsConstructor
@@ -12,12 +11,11 @@ import java.io.Serializable;
 @ToString
 @Entity
 @Table(name = "detail_work")
-public class TDetailWork implements Serializable {
-
+public class TDetailWork {
     @Id
     @Column(name = "id_user")
     private Long idUser;
-    @Id
+    @PrimaryKeyJoinColumn(name = "id_work")
     @Column(name = "id_work")
     private Long idWork;
     private String description;
@@ -25,13 +23,5 @@ public class TDetailWork implements Serializable {
     @OneToOne
     @JoinColumn(name = "id_detail_type", nullable = false)
     private TDetailType detailType;
-
-    @ManyToOne(cascade = CascadeType.ALL)
-    @JoinColumn(name = "id_user", nullable = false)
-    private TUser user;
-
-    @ManyToOne(cascade = CascadeType.ALL)
-    @JoinColumn(name = "id_work", nullable = false)
-    private TWorks works;
 
 }
