+++ src/main/java/com/pmt/tool/entity/TRole.java
@@ -1,0 +1,23 @@
+package com.pmt.tool.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import javax.persistence.*;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+@Entity
+@Table(name = "role")
+public class TRole {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id_role")
+    private Long idRole;
+    @Column(nullable = false)
+    private String nameRole;
+}
