+++ src/main/java/com/pmt/tool/repositories/RoleRepository.java
@@ -1,0 +1,9 @@
+package com.pmt.tool.repositories;
+
+import com.pmt.tool.entity.TRole;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface RoleRepository extends JpaRepository<TRole, Long> {
+}
