+++ src/main/resources/application.properties
@@ -1,14 +1,18 @@
+server.port=8085
 ## Hibernate Properties
 # The SQL dialect makes Hibernate generate better SQL for the chosen database
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5InnoDBDialect
 # Hibernate ddl auto (create, create-drop, validate, update)
-logging.level.org.hibernate=ERROR
-spring.datasource.url=jdbc:mysql://localhost:3306/tools?createDatabaseIfNotExist=true&maxAllowedPacket=524288000
+spring.datasource.url=jdbc:mysql://localhost:3308/tools\
+  ?createDatabaseIfNotExist=true\
+  &maxAllowedPacket=1073741824
 spring.datasource.username=root
-spring.datasource.password=password
+spring.datasource.password=Abc@123456789
 spring.datasource.driverClassName=com.mysql.cj.jdbc.Driver
 spring.jpa.show-sql=true
 spring.jpa.hibernate.ddl-auto=update
+spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
 ## MULTIPART (MultipartProperties)
 # Enable multipart uploads
 spring.servlet.multipart.enabled=true
